feat + fix: ammount field in cart is now int64 and RemoveItemInCart implemented for ValkeyCartManager

+++ main.go
@@ -3,6 +3,6 @@ package main
 import "fmt"
 
 func main() {
-    x := Cart{2626, 1727, map[uint64]int{}}
+    x := Cart{2626, 1727, map[uint64]int64{}}
     fmt.Println(x)
 }
